add users voter

+++ src/Controller/Admin/TrickManageController.php
@@ -75,6 +75,7 @@ class TrickManageController extends AbstractController
      */
     public function edit(Trick $trick, Request $request)
     {
+        $this->denyAccessUnlessGranted('TRICK_EDIT', $trick);
         $form = $this->createForm(TrickType::class, $trick);
         $form->handleRequest($request);
 
@@ -96,6 +97,7 @@ class TrickManageController extends AbstractController
      */
     public function delete(Trick $trick, Request $request)
     {
+        $this->denyAccessUnlessGranted('TRICK_DELETE', $trick);
         if($this->isCsrfTokenValid('delete'.$trick->getId(), $request->get('_token'))){
             $em = $this->getDoctrine()->getManager();
             $em->remove($trick);
